+++ tests/testthat/programs/sample26.R
@@ -1,0 +1,4 @@
+#% Assign variables
+#%let x <- 1
+#%let y <- %nrstr(&x.)
+print("The value of &y is &x")