+++ tests/testthat/programs/test27_mod.R
@@ -1,0 +1,6 @@
+a <- "Here is Hello a regular"
+b <- "Here is &x. a protected"
+c <- "Here is &x. multiple &y. protected"
+d <- "Here is Hello mixed &x."
+e <- "Here is &x. with unresolved reference"
+